// src/ports.d.ts
/* This file was generated by github.com/ronanyeah/elm-port-gen */

interface ElmApp {
  ports: Ports;
}

interface Ports {
  log: PortOut<string>;
  copy: PortOut<string>;
  deauthorize: PortOut<number>;
  refreshPrice: PortOut<string>;
  authorizeSeeds: PortOut<null>;
  fetchSeed: PortOut<SeedAuth>;
  submitTx: PortOut<{
    amount: number;
    seed: Seed;
    recipient: string;
  }>;
  seedCb: PortIn<Seed | null>;
  priceCb: PortIn<number | null>;
  sigCb: PortIn<string | null>;
  authedCb: PortIn<SeedAuth[]>;
}

interface PortOut<T> {
  subscribe: (_: (_: T) => void) => void;
}

interface PortIn<T> {
  send: (_: T) => void;
}

interface Seed {
  pubkey: string;
  pubkeyBytes: number[];
  balance: number;
  auth: SeedAuth;
}

interface SeedAuth {
  name: string;
  authToken: number;
}

export { ElmApp, Seed, SeedAuth };